Disassociate type families, drop magic typeclass and TH
We need to separate the `ResourceData` and `StoredResourceData`
families from the `Resource` typeclass, so we can provide these
instances in API packages that don't include handler implementations.

Once that's done, there's no need to have the type class at all, and
then the need for Template Haskell also falls away.